# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Personal Finance API - Gastos
  description: API para gestión de gastos en aplicación de finanzas personales
  version: 1.0.0
  contact:
    name: Personal Finance API
    url: http://localhost:3030
servers:
  - url: http://localhost:3030/api
    description: Development server
...
components:
  schemas:
    StandardResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indica si la operación fue exitosa
        data:
          description: Datos de la respuesta
        meta:
          type: object
          properties:
            total:
              type: integer
              description: Total de elementos
            type:
              type: string
              description: Tipo de respuesta
              enum: [collection, single]
            pagination:
              $ref: '#/components/schemas/Pagination'
...
    ValidationError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error de validación"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Campo que falló la validación
              message:
                type: string
                description: Mensaje de error específico
              value:
                description: Valor que causó el error
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Número de elementos por página
        offset:
          type: integer  
          description: Número de elementos saltados
        hasNext:
          type: boolean
          description: Indica si hay página siguiente
        hasPrev:
          type: boolean
          description: Indica si hay página anterior
...
    GastoUnico:
      type: object
      properties:
        id:
          type: integer
          description: ID único del gasto único
          example: 789
        descripcion:
          type: string
          minLength: 3
          maxLength: 255
          description: Descripción del gasto
          example: "Compra única de equipo"
        monto:
          type: number
          minimum: 0.01
          description: Monto del gasto
          example: 2500.75
        fecha:
          type: string
          format: date
          description: Fecha del gasto (no puede ser futura)
          example: "2024-01-15"
        categoria_gasto_id:
          type: integer
          description: ID de la categoría del gasto
          example: 8
        importancia_gasto_id:
          type: integer
          description: ID de la importancia del gasto
          example: 1
        tipo_pago_id:
          type: integer
          description: ID del tipo de pago
          example: 3
        tarjeta_id:
          type: integer
          nullable: true
          description: ID de la tarjeta (opcional)
          example: 1
        procesado:
          type: boolean
          description: Indica si ya fue procesado para generar gasto real
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        categoria:
          $ref: '#/components/schemas/Categoria'
        importancia:
          $ref: '#/components/schemas/Importancia'
        tipoPago:
          $ref: '#/components/schemas/TipoPago'
        tarjeta:
          $ref: '#/components/schemas/Tarjeta'
...
    Categoria:
      type: object
      properties:
        id:
          type: integer
        nombre_categoria:
          type: string
    Importancia:
      type: object
      properties:
        id:
          type: integer
        nombre_importancia:
          type: string
    TipoPago:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        permite_cuotas:
          type: boolean
    Tarjeta:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        tipo:
          type: string
        banco:
          type: string
        dia_cierre:
          type: integer
        dia_vencimiento:
          type: integer
        permite_cuotas:
          type: boolean
...
paths:
...
  /gastos-unicos:
    get:
      tags:
        - gastos-unicos
      summary: Obtener gastos únicos con filtros
      description: |
        Obtiene gastos únicos aplicando filtros opcionales con paginación inteligente.
        
        **Comportamiento de paginación:**
        - Sin `limit`: Devuelve todos los resultados filtrados
        - Con `limit`: Devuelve respuesta paginada con metadata de navegación
      parameters:
        - name: categoria_gasto_id
          in: query
          schema:
            type: integer
          description: Filtrar por categoría de gasto
        - name: importancia_gasto_id
          in: query
          schema:
            type: integer
          description: Filtrar por importancia
        - name: tipo_pago_id
          in: query
          schema:
            type: integer
          description: Filtrar por tipo de pago
        - name: tarjeta_id
          in: query
          schema:
            type: integer
          description: Filtrar por tarjeta
        - name: fecha_desde
          in: query
          schema:
            type: string
            format: date
          description: Fecha de inicio del rango
        - name: fecha_hasta
          in: query
          schema:
            type: string
            format: date
          description: Fecha de fin del rango
        - name: monto_min
          in: query
          schema:
            type: number
          description: Monto mínimo
        - name: monto_max
          in: query
          schema:
            type: number
          description: Monto máximo
        - name: procesado
          in: query
          schema:
            type: boolean
          description: Filtrar por estado procesado
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Número máximo de resultados (activa paginación)
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Número de resultados a omitir
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [fecha, monto, descripcion, createdAt, updatedAt]
            default: fecha
          description: Campo por el cual ordenar
        - name: orderDirection
          in: query
          schema:
            type: string
            enum: [ASC, DESC]
            default: DESC
          description: Dirección del ordenamiento
      responses:
        '200':
          description: Lista de gastos únicos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/GastoUnico'

    post:
      tags:
        - gastos-unicos
      summary: Crear gasto único
      description: |
        Crea un nuevo gasto único. Al crearse, se programa automáticamente
        para generar un gasto real en la tabla principal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - descripcion
                - monto
                - fecha
                - categoria_gasto_id
                - importancia_gasto_id
                - tipo_pago_id
              properties:
                descripcion:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: "Compra de equipamiento"
                monto:
                  type: number
                  minimum: 0.01
                  example: 2500.50
                fecha:
                  type: string
                  format: date
                  description: Fecha del gasto (no puede ser futura)
                  example: "2024-01-15"
                categoria_gasto_id:
                  type: integer
                  minimum: 1
                  example: 8
                importancia_gasto_id:
                  type: integer
                  minimum: 1
                  example: 2
                tipo_pago_id:
                  type: integer
                  minimum: 1
                  example: 3
                tarjeta_id:
                  type: integer
                  minimum: 1
                  nullable: true
                  description: Requerido para pagos con tarjeta
                  example: 2
            example:
              descripcion: "Compra de equipamiento"
              monto: 2500.50
              fecha: "2024-01-15"
              categoria_gasto_id: 8
              importancia_gasto_id: 2
              tipo_pago_id: 3
              tarjeta_id: 2
      responses:
        '201':
          description: Gasto único creado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/StandardResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GastoUnico'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...